+++ frontend/tests/components/charts/VerticalBarChart.test.tsx
@@ -1,87 +1,16 @@
 import { render, screen, waitFor } from "@testing-library/react";
 import VerticalBarChart from "@components/charts/VerticalBarChart.tsx";
 import { beforeAll } from "vitest";
+import {renderMocksForResponsiveContainer} from "../../test-utils.tsx";
 
 interface DataPoint {
   name: string; // Y-Achse Label (z.B. Themenname, Modulname)
   value: number; // x-Achse Wert (z.B. Stunden)
 }
 
-const originalGetBoundingClientRect =
-  HTMLElement.prototype.getBoundingClientRect;
-
 beforeAll(() => {
-  // Make getBoundingClientRect return non-zero for the recharts wrapper
-  HTMLElement.prototype.getBoundingClientRect = function () {
-    if (
-      this.classList &&
-      this.classList.contains("recharts-responsive-container")
-    ) {
-      return {
-        width: 800,
-        height: 300,
-        top: 0,
-        left: 0,
-        right: 800,
-        bottom: 300,
-        x: 0,
-        y: 0,
-        toJSON: () => {},
-      } as DOMRect;
-    }
-    return (
-      originalGetBoundingClientRect?.call(this) ?? {
-        width: 0,
-        height: 0,
-        top: 0,
-        left: 0,
-        right: 0,
-        bottom: 0,
-        x: 0,
-        y: 0,
-        toJSON: () => {},
-      }
-    );
-  };
-
-  // offsetWidth/offsetHeight positive values
-  Object.defineProperty(HTMLElement.prototype, "offsetWidth", {
-    configurable: true,
-    get() {
-      return 800;
-    },
-  });
-  Object.defineProperty(HTMLElement.prototype, "offsetHeight", {
-    configurable: true,
-    get() {
-      return 300;
-    },
-  });
-
-  // requestAnimationFrame polyfill to run callbacks
-  (global as any).requestAnimationFrame = (cb: FrameRequestCallback) =>
-    setTimeout(cb, 0);
-
-  // ResizeObserver mock that immediately notifies with current rect
-  class MockResizeObserver {
-    cb: any;
-    constructor(cb: any) {
-      this.cb = cb;
-    }
-    observe(target: Element) {
-      const rect = (target as any).getBoundingClientRect?.() ?? {
-        width: 800,
-        height: 300,
-      };
-      this.cb([
-        { target, contentRect: { width: rect.width, height: rect.height } },
-      ]);
-    }
-    unobserve() {}
-    disconnect() {}
-  }
-  (global as any).ResizeObserver = MockResizeObserver;
-});
+    renderMocksForResponsiveContainer()
+})
 
 describe("VerticalBarChart", async () => {
   // setup testing data
